Removes some compilation warnings
