+++ project/plugins.sbt
@@ -1,3 +1,2 @@
-resolvers += Resolver.sonatypeRepo("snapshots")
-addSbtPlugin("com.47deg" % "sbt-org-policies" % "0.4.3-SNAPSHOT")
+addSbtPlugin("com.47deg" % "sbt-org-policies" % "0.4.3")
 libraryDependencies += "org.scala-sbt" % "scripted-plugin" % sbtVersion.value
